backend/requirements.txt
absl-py
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
alembic
altair
amqp
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
Arpeggio
arrow
arxiv
asttokens
async-lru
attrs
Authlib
autogui
babel
backoff
bcrypt
beautifulsoup4
billiard
black
bleach
blinker
build
CacheControl
cachetools
celery
certifi
cffi
chardet
charset-normalizer
chromadb
cleo
click
click-didyoumean
click-plugins
click-repl
clr-loader
colorama
coloredlogs
comm
contourpy
crashtest
cryptography
cycler
dataclasses-json
debugpy
decorator
defusedxml
dill
distlib
distro
dnspython
dulwich
durationpy
ecdsa
email_validator
emoji
et-xmlfile
exceptiongroup
executing
fake-useragent
fastapi
fastjsonschema
fastmcp
feedparser
filelock
filetype
findpython
Flask
flask_wrappers
flatbuffers
fonttools
fpdf
fqdn
frozendict
frozenlist
fsspec
git-filter-repo
gitdb
GitPython
google-auth
google-auth-oauthlib
googleapis-common-protos
greenlet
groq
grpcio
gspread
h11
html5lib
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
iniconfig
installer
ipykernel
ipython
isoduration
isort
itsdangerous
jaraco.classes
jaraco.context
jaraco.functools
jax
jaxlib
jedi
Jinja2
jiter
joblib
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
keyring
kiwisolver
kombu
kubernetes
langchain
langchain-community
langchain-core
langchain-experimental
langchain-groq
langchain-openai
langchain-text-splitters
langchainhub
langdetect
langsmith
llvmlite
loguru
lxml
mailchimp-transactional
Mako
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mcp
mdurl
mediapipe
mistune
ml_dtypes
mmh3
more-itertools
motor
MouseInfo
mplcyberpunk
mpmath
msgpack
multidict
multitasking
mypy
mypy-extensions
mysql-connector-python
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
nodeenv
notebook
notebook_shim
numba
numpy
oauth2client
oauthclient
oauthlib
odf
odfpy
olefile
onnxruntime
openai
openai-whisper
openapi-pydantic
opencv-contrib-python
opencv-python
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opt_einsum
orjson
outcome
overrides
packaging
pandas
pandocfilters
parso
parver
passlib
pathspec
pbs-installer
peewee
pillow
pkginfo
platformdirs
playwright
plotly
pluggy
poetry
poetry-core
posthog
prometheus_client
prompt_toolkit
propcache
protobuf
psutil
psycopg2-binary
pulumi
pulumi_gcp
pure_eval
pyarrow
pyasn1
pyasn1_modules
PyAutoGUI
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
pyee
PyGetWindow
Pygments
pymongo
PyMsgBox
PyMySQL
pynput
pyparsing
pypdf
pyperclip
PyPika
pyproject_hooks
pyreadline3
PyRect
pyright
PyScreeze
PySocks
pytest
pytest-html
pytest-metadata
pytest-timeout
python-dateutil
python-docx
python-dotenv
python-iso639
python-jose
python-json-logger
python-magic
python-multipart
python-oxmsg
pythonnet
pytweening
pytz
pywin32
pywin32-ctypes
pywinpty
PyYAML
pyzmq
RapidFuzz
redis
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rpds-py
rsa
scipy
selenium
semver
Send2Trash
sentencepiece
setuptools
sgmllib3k
shellingham
six
smmap
sniffio
sortedcontainers
sounddevice
soundfile
soupsieve
SQLAlchemy
SQLAlchemy-Utils
sse-starlette
stack-data
StackAPI
starlette
streamlit
sympy
tabulate
tenacity
terminado
tiktoken
tinycss2
tk
tokenizers
toml
tomlkit
torch
tornado
tqdm
traitlets
trio
trio-websocket
trove-classifiers
typer
types-python-dateutil
types-requests
typing-inspect
typing-inspection
typing_extensions
tzdata
unstructured
unstructured-client
uri-template
urllib3
uv
uvicorn
vboxapi
vine
virtualenv
watchdog
watchfiles
wcwidth
webcolors
webencodings
websocket-client
websockets
Werkzeug
wheel
wikipedia
win32_setctime
wrapt
wsproto
yarl
yfinance
youtube-transcript-api
zipp
zstandard
